# public/openapi.yaml
openapi: 3.0.3
info:
  title: Gerobaks API
  description: |
    # 🚛 Gerobaks Waste Management API
    
    RESTful API Backend untuk sistem manajemen pengumpulan sampah Gerobaks.
    
    ## 🌟 Features
    - 🔐 Authentication & Authorization (Laravel Sanctum)
    - 📅 Schedule Management (Mobile & Standard Format)
    - 📍 Real-time GPS Tracking
    - 🛒 Order Management
    - 💰 Payment & Balance System
    - ⭐ Rating & Reviews
    - 💬 Chat System
    - 🔔 Notifications
    - 📊 Dashboard & Analytics
    - 📝 Reports & Feedback
    - 🎯 Subscription Plans
    - ⚙️ Settings Management
    - 👨‍💼 Admin Management
    
    ## 📖 Repository
    - **GitHub**: [https://github.com/fk0u/gerobackend](https://github.com/fk0u/gerobackend)
    - **Developer**: [@fk0u](https://github.com/fk0u)
    
    ## 🔐 Authentication
    API menggunakan **Laravel Sanctum** untuk autentikasi.
    
    **Flow:**
    1. POST /api/login → Get token
    2. Use token in header: `Authorization: Bearer {token}`
    3. POST /api/auth/logout → Revoke token
    
  version: 1.0.0
  contact:
    name: fk0u
    url: https://github.com/fk0u
    email: support@gerobaks.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://gerobaks.dumeg.com/api
    description: Production Server
  - url: https://staging-gerobaks.dumeg.com/api
    description: Staging Server
  - url: http://127.0.0.1:8000/api
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Login untuk mendapatkan token:
        ```
        Authorization: Bearer YOUR_TOKEN_HERE
        ```
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [end_user, mitra, admin]
        phone:
          type: string
        address:
          type: string
        balance:
          type: number
        created_at:
          type: string
          format: date-time
...
paths:
...
  /register:
    post:
      tags: [Authentication]
      summary: Register New User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, password_confirmation, role]
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                  minLength: 8
                password_confirmation:
                  type: string
                phone:
                  type: string
                role:
                  type: string
                  enum: [end_user, mitra, admin]
      responses:
        '201':
          description: User registered
        '422':
          description: Validation error
  /login:
    post:
      tags: [Authentication]
      summary: Login User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: daffa@gmail.com
                password:
                  type: string
                  example: password123
            examples:
              end_user:
                summary: End User Login
                value:
                  email: daffa@gmail.com
                  password: password123
              mitra:
                summary: Mitra Login
                value:
                  email: mitra@gerobaks.com
                  password: password123
              admin:
                summary: Admin Login
                value:
                  email: admin@gerobaks.com
                  password: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
        '401':
          description: Invalid credentials
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get Current User
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthenticated
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out
...